+++ libraries/utils.rb
@@ -13,30 +13,36 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require "net/http"
-require "uri"
-require "json"
+require 'net/http'
+require 'uri'
+require 'json'
 
+if RUBY_PLATFORM =~ /mswin|mingw32|windows/
+  ::Chef::Recipe.send(:include, Windows::RegistryHelper)
+end
+
+#
+# Utilities used by various arcgis cookbook resources.
+#
 module Utils
-
   MAX_RETRIES = 30
   SLEEP_TIME = 10.0
 
-  def Utils.wait_until_url_available(url)
+  def self.wait_until_url_available(url)
     uri = URI.parse(url)
 
     http = Net::HTTP.new(uri.host, uri.port)
     http.read_timeout = 600
 
-    if uri.scheme == "https"
+    if uri.scheme == 'https'
       http.use_ssl = true
       http.verify_mode = OpenSSL::SSL::VERIFY_NONE
     end
 
-    for i in 1..MAX_RETRIES
+    MAX_RETRIES.times do
       begin
         request = Net::HTTP::Get.new(uri.request_uri)
-        request.add_field("Accept-Language", "en-US")
+        request.add_field('Accept-Language', 'en-US')
         response = http.request(request)
 
         if response.code.to_i < 400
@@ -50,37 +56,27 @@ module Utils
       end
 
       sleep(SLEEP_TIME)
-    end 
-  end
-  
-  def Utils.load_properties(properties_filename)
-    properties = {}
-    File.open(properties_filename, 'r') do |properties_file|
-      properties_file.read.each_line do |line|
-        line.strip!
-        if (line[0] != ?# and line[0] != ?=)
-          i = line.index('=')
-          if (i)
-            properties[line[0..i - 1].strip] = line[i + 1..-1].strip
-          else
-            properties[line] = ''
-          end
-        end
-      end      
     end
-    properties
   end
-  
-  def Utils.configure_ssl(pfx_file, keystore_pass, appid)
+
+  def self.configure_ssl(pfx_file, keystore_pass, appid)
     pkcs12 = OpenSSL::PKCS12.new(::File.binread(pfx_file), keystore_pass)
     certhash = Digest::SHA1.hexdigest(pkcs12.certificate.to_der)
-    
+
     cmd = Mixlib::ShellOut.new("certutil -f -p \"#{keystore_pass}\" -importpfx \"#{pfx_file}\"")
     cmd.run_command
     cmd.error!
-    
+
+    cmd = Mixlib::ShellOut.new("netsh http delete sslcert ipport=0.0.0.0:443")
+    cmd.run_command
+
     cmd = Mixlib::ShellOut.new("netsh http add sslcert ipport=0.0.0.0:443 certhash=#{certhash} appid=#{appid}")
     cmd.run_command
     cmd.error!
+  end
+
+  def self.product_installed?(product_code)
+    Registry.key_exists?('HKEY_LOCAL_MACHINE\\SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\Uninstall\\' + product_code) ||
+      Registry.key_exists?('HKEY_LOCAL_MACHINE\\SOFTWARE\\Wow6432Node\\Microsoft\\Windows\\CurrentVersion\\Uninstall\\' + product_code)
   end
 end