+++ recipes/iptables.rb
@@ -20,4 +20,3 @@
 iptables_rule 'iptables' do
   action :enable
 end
-