+++ resources/licensemanager.rb
@@ -17,9 +17,10 @@
 # limitations under the License.
 #
 
-actions :install
+actions :system, :install, :uninstall
 
 attribute :setup, :kind_of => String
+attribute :product_code, :kind_of => String
 attribute :install_dir, :kind_of => String
 attribute :python_dir, :kind_of => String
 attribute :run_as_user, :kind_of => String