+++ resources/server.rb
@@ -17,9 +17,10 @@
 # limitations under the License.
 #
 
-actions :install, :authorize, :create_site, :join_site, :join_cluster, :enable_ssl, :register_database, :federate 
+actions :system, :install, :uninstall, :authorize, :create_site, :join_site, :join_cluster, :configure_https, :register_database, :federate
 
 attribute :setup, :kind_of => String
+attribute :product_code, :kind_of => String
 attribute :install_dir, :kind_of => String
 attribute :python_dir, :kind_of => String
 attribute :run_as_user, :kind_of => String
@@ -27,6 +28,9 @@ attribute :run_as_password, :kind_of => String
 attribute :server_url, :kind_of => String
 attribute :server_local_url, :kind_of => String
 attribute :primary_server_url, :kind_of => String
+attribute :keystore_file, :kind_of => String
+attribute :keystore_password, :kind_of => String
+attribute :cert_alias, :kind_of => String
 attribute :cluster, :kind_of => String
 attribute :server_id, :kind_of => String
 attribute :secret_key, :kind_of => String
@@ -36,11 +40,15 @@ attribute :portal_password, :kind_of => String
 attribute :username, :kind_of => String
 attribute :password, :kind_of => String
 attribute :server_directories_root, :kind_of => String
+attribute :config_store_connection_string, :kind_of => String 
+attribute :config_store_connection_secret, :kind_of => String
+attribute :config_store_type, :kind_of => String
 attribute :authorization_file, :kind_of => String
 attribute :authorization_file_version, :kind_of => String
 attribute :data_item_path, :kind_of => String
 attribute :connection_string, :kind_of => String
 attribute :is_managed, :kind_of => [TrueClass, FalseClass], :default => false
+attribute :system_properties, :kind_of => Hash, :default => {}
 
 def initialize(*args)
   super