+++ CHANGELOG.md
@@ -21,5 +21,12 @@ This file is used to list changes made in each version of the arcgis cookbook.
 - Added ArcGIS GeoEvent Extension for Server resource
 - Default value of attribute node['data_store']['preferredidentifier'] set to 'hostname' instead of 'ip' 
 
-        
-
+2.2.0
+-----
+- This version supports only ArcGIS 10.4 release software. (Use 1.1.3 version with ArcGIS software 10.3.1 release.)
+- Added 'all_uninstalled' and 'cleanup' recipes
+- Added 'arcgis' top level key to all arcgis cookbook attributes
+- Added support for ArcGIS Pro
+- Using 10.4 federation and HA workflows 
+- Data Store types made configurable (see node['arcgis']['data_store']['types'] attribute)
+- Added :uninstall action to geoevent resource